feature(slim-api): moved `chain` commands back to top-level
