+++ packages/groqd/src/commands/order.ts
@@ -1,8 +1,8 @@
-import { GroqBuilderChain } from "../../groq-builder";
-import { ResultItem } from "../../types/result-types";
-import { Expressions } from "../../types/groq-expressions";
+import { GroqBuilderChain } from "../groq-builder";
+import { ResultItem } from "../types/result-types";
+import { Expressions } from "../types/groq-expressions";
 
-declare module "../../groq-builder" {
+declare module "../groq-builder" {
   export interface GroqBuilderChain<TResult, TQueryConfig> {
     /**
      * Orders the results by the keys specified
